HBP-87. Move tests from core at top-level.
