+++ tvb_multiscale/tests/core/DMF2014/matlab_rww_dfun.m
