+++ tvb_multiscale/tests/core/DMF2014/phie.m
