+++ tvb_multiscale/tests/core/test_model_rwwei.py
